requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
firecrawl-py==0.0.16
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
openai==1.3.0

